Refactor dragon.js
